requirements.txt
#py-cord
git+https://github.com/Pycord-Development/pycord.git
python-dotenv
openai
emoji
# Google api
google-api-python-client
google-cloud-vision
tiktoken
bs4
discord-oauth2.py
black
orjson # for speed
simpleeval